Move esbuild plugins handler in its file
